Complete code refractoring and Blackjack logic
Code refractored from scratch. Features that were broken have been worked on (such as the blackjack game being barebones).
Updated the entire structure of the server, as implementing storage or new HTMLs was a mess.
Templates are partially complete, still work in progress, blackjack is the main focus. Made a simple hub so the user will be able to click on blackjack or card-game (earlier assignment). Next step is implementing storage for the blackjack game, and updating the css and templates.
Massive commit due to much of the code not working as intended, so a lot of changes were made to ensure functionality, and future structure for missing functionality.